Partially restored "Moved from browser-builder to application-builder"
The new application-builder will not be used for the time being. The new builder seems to use an old version of Vite, which has problems when dots are in the end-section of the current URL (see https://github.com/vitejs/vite/issues/2415 ).

Accessing a versioned tool-path causes a 404. This could be worked around by adding a trailing slash, but this is automatically removed upon the route being loaded.
* http://localhost:4200/minecraft_toolbox/loot-table-randomizer/1.20.0 -> 404
* http://localhost:4200/minecraft_toolbox/loot-table-randomizer/1.20.0/ -> Loads like normal

This isn't a problem for the end-product, but is very annoying during development.

This commit partially restores some of the changed behaviour to make the transition in the future a bit easier.